pyproject.toml
[tool.poetry]
name = "bigfoot"
version = "0.1.0"
description = "The missing (big) schema manager for Google BigQuery."
authors = ["Logan Hennessy <logan@gradient.com>"]

[tool.poetry.dependencies]
python = "^3.9,<3.11"
typer = "^0.4.0"
google-cloud-bigquery = "^2.34.0"
google-api-python-client = "^2.37.0"
google-auth-oauthlib = "^0.4.6"
pydantic = "^1.9.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
